added Readme.md and few minor changes

+++ src/main/java/com/todolist/service/Impl/ToDoListServiceImpl.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ToDoListServiceImpl implements ToDoListService {
@@ -28,7 +29,7 @@ public class ToDoListServiceImpl implements ToDoListService {
     }
 
     @Override
-    public String addTask(ToDoListDto toDoListDto) {
+    public ToDoListDto addTask(ToDoListDto toDoListDto) {
         try {
             ToDoList toDoLists = toDoListDto.convertDtoToEntity(toDoListDto);
             toDoListRepository.save(toDoLists);
@@ -36,18 +37,28 @@ public class ToDoListServiceImpl implements ToDoListService {
         catch (Exception e){
             e.printStackTrace();
         }
-        return "Task added successfully";
+        return toDoListDto;
     }
 
     @Override
     public ToDoListDto modifyTask(ToDoListDto updatedToDoList) {
-        ToDoList updatedToDoListResult = toDoListRepository.save(ToDoListDto.convertDtoToEntity(updatedToDoList));
-        ToDoListDto toDoListDtoResult =  ToDoList.convertEntityToDto(updatedToDoListResult);
+        Optional<ToDoList> toDoList = toDoListRepository.findById(updatedToDoList.getId());
+        ToDoListDto toDoListDtoResult;
+        if (toDoList.isEmpty()) {
+            throw new RuntimeException("Id not found in tasks");
+        } else {
+            ToDoList updatedToDoListResult = toDoListRepository.save(ToDoListDto.convertDtoToEntity(updatedToDoList));
+            toDoListDtoResult = ToDoList.convertEntityToDto(updatedToDoListResult);
+        }
         return toDoListDtoResult;
     }
 
     @Override
     public String deleteTask(Integer id) {
+        Optional<ToDoList> toDoList = toDoListRepository.findById(id);
+        if (toDoList.isEmpty()) {
+            throw new RuntimeException("Id not found in tasks");
+        }
         toDoListRepository.deleteById(id);
         return "Task Deleted Successfully";
     }
